jobs-service/requirements.txt
annotated-types==0.7.0
azure-functions==1.23.0
dnspython==2.7.0
MarkupSafe==3.0.2
motor==3.7.0
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
typing-inspection==0.4.0
typing_extensions==4.13.2
Werkzeug==3.1.3
